requirements.txt
requests>=2.28.0
pyyaml>=6.0
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.2
